Move downsample to pool folder from data_movement folder
